Struct rust-conway::engine::GrifLife

pub struct GrifLife {
    // some fields omitted
}

The main implementation of Conway's game of life. This engine utilizes a list of updates to track interesting areas in the world. Only cells adjacent to cells that updated during the last generation are evaluated for a new state

Methods

impl GrifLife

fn new(world: Box<World + 'static>) -> GrifLife

Create a new instance of the engine, this should be used on a world with an initial setup of cells.