Trait rust-conway::engine::ConwayEngine

pub trait ConwayEngine {
    fn next_generation(&mut self);
    fn world_ref<'w>(&'w self) -> &'w World;
    fn world_ref_mut<'w>(&'w mut self) -> &'w mut World;
    fn set_cell(&mut self, x: isize, y: isize);
}

Required Methods

fn next_generation(&mut self)

Computes the next generation of cells, works on an internally held world::World object

fn world_ref<'w>(&'w self) -> &'w World

Return an immutable reference to the internally held world object. This is useful for getting access to the world so that It can be printed

fn world_ref_mut<'w>(&'w mut self) -> &'w mut World

Return a mutable reference to the internally held world object.

fn set_cell(&mut self, x: isize, y: isize)

Correct way to set the cell in the engine so we can know about the addition