Trait rust-conway::world::World

pub trait World {
    fn get_cell(&self, x: isize, y: isize) -> State;
    fn set_cell(&mut self, x: isize, y: isize);
    fn kill_cell(&mut self, x: isize, y: isize);
    fn iter(&self) -> Iter<(isize, isize), State>;

    fn num_adjacent(&self, x: isize, y: isize) -> isize { ... }
}

Required Methods

fn get_cell(&self, x: isize, y: isize) -> State

fn set_cell(&mut self, x: isize, y: isize)

fn kill_cell(&mut self, x: isize, y: isize)

fn iter(&self) -> Iter<(isize, isize), State>

Returns a HashMap iterator of live cells. This will stay in this format for the forseeable future. Returning a generic iterator adds too much overhead to the iter() function

Provided Methods

fn num_adjacent(&self, x: isize, y: isize) -> isize